fix: tune dependency as polkadot-sdk

+++ node/src/benchmarking.rs
@@ -121,7 +121,7 @@ pub fn create_benchmark_extrinsic(
 			period,
 			best_block.saturated_into(),
 		)),
-		frame_system::CheckNonce::<runtime::Runtime>::from(nonce),
+		frame_system::CheckNonce::<runtime::Runtime>::from(nonce as u64),
 		frame_system::CheckWeight::<runtime::Runtime>::new(),
 		pallet_transaction_payment::ChargeTransactionPayment::<runtime::Runtime>::from(0),
 	);
